+++ EntityLib/include/EntityLib/Property.h
@@ -8,7 +8,7 @@ namespace Ent
     struct ENTLIB_DLLEXPORT Property
     {
     public:
-        using Key = std::variant<std::string, size_t>;
+        using Key = std::variant<std::string, int64_t>;
 
         Property(Property const& _other)
             : m_self(_other.m_self == nullptr ? nullptr : _other.m_self->sharedFromThis())
